+++ CHANGELOG.md
@@ -8,6 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 ### Changed
 ### Fixed
+- Genes API endpoint to return a json formatted error when request is malformed
 
 ## [4.43.1]
 ### Added